Changed page's name
